requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
contourpy==1.2.1
crispy-bootstrap4==2024.1
cryptography==45.0.4
cycler==0.12.1
distro==1.9.0
Django==5.1
django-crispy-forms==2.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
fonttools==4.53.1
frozenlist==1.4.1
geopandas==1.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
kiwisolver==1.4.5
matplotlib==3.9.2
multidict==6.0.5
numpy==2.1.0
openai==0.28.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyogrio==0.9.0
pyparsing==3.1.4
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.1
requests==2.32.3
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
